refactor existing code in ./mindmeld/models
