+++ mindmeld/models/features/entity_features.py
@@ -12,8 +12,8 @@
 # limitations under the License.
 
 """This module contains feature extractors for entities"""
-from .helpers import GAZETTEER_RSC, get_ngram, register_entity_feature, requires
-from ..core import QueryEntity, Query, NestedEntity
+from mindmeld.models.helpers import GAZETTEER_RSC, get_ngram, register_entity_feature, requires
+from mindmeld.core import QueryEntity, Query, NestedEntity
 
 
 @register_entity_feature(feature_name="in-gaz")
